Add PusherAssembly
