+++ Sources/Framework/Dependency Injection/Assembler+Assemblies.swift
@@ -10,7 +10,7 @@ public extension Assembler {
 
     static var defaultAssemblies: [Assembly] {
         return [MainAssembly(), PinEntryAssembly(), AnswerAssembly(),
-                CommentAssembly(), ValidatorAssembly(), APIAssembly(), ServiceHelpersAssembly(), CommentsServicesAssembly()]
+                CommentAssembly(), ValidatorAssembly(), APIAssembly(), ServiceHelpersAssembly(), CommentsServicesAssembly(), PusherAssembly()]
     }
 
 }
